Main Loop
This is the main part of the code. Here from the momentom equation u and v components of particle velocity in each time step will be calculated. Therefore, the new position of particle can be obtained. This loop continues until the particle reaches to the end of channel or gets too close to the wall.